Install Lazygit checks

+++ install_lazygit.sh
@@ -10,17 +10,19 @@ else
     . ./aliases/rlwrap_scripts.sh
 fi
 
-reade -Q "GREEN" -i "y" -p "Install lazygit? (Git gui) [Y/n]: " "y n" nstll
-if [ "$nstll" == "y" ]; then
-    if [ $distro == "Arch" ] || [ $distro_base == "Arch" ]; then
-        yes | sudo pacman -Su lazygit
-    elif [ $distro == "Debian" ] || [ $distro_base == "Debian" ]; then
-        yes | sudo apt update
-        yes | sudo apt install lazygit
+if ! type lazygit &> /dev/null; then
+    reade -Q "GREEN" -i "y" -p "Install lazygit? (Git gui) [Y/n]: " "y n" nstll
+    if [ "$nstll" == "y" ]; then
+        if [ $distro == "Arch" ] || [ $distro_base == "Arch" ]; then
+            yes | sudo pacman -Su lazygit
+        elif [ $distro == "Debian" ] || [ $distro_base == "Debian" ]; then
+            yes | sudo apt update
+            yes | sudo apt install lazygit
+        fi
+        
     fi
-    
+    unset nstll
 fi
-unset nstll
 
 if ! type copy-to &> /dev/null; then
     reade -Q "GREEN" -i "y" -p "Install copy-to? [Y/n]: " "y n" cpcnf;
@@ -33,16 +35,18 @@ if ! type copy-to &> /dev/null; then
     fi
 fi
 
-reade -Q "GREEN" -i "y" -p "Set up an alias so copy-to does a 'run all' when starting up lazygit? [Y/n]: " "y n" nstll
-if [ "$nstll" == "y" ]; then
-    if ! test -f checks/check_aliases_dir.sh; then
-        eval "$(curl -fsSL https://raw.githubusercontent.com/excited-bore/dotfiles/main/checks/check_aliases_dir.sh)" 
-    else
-       . ./checks/check_aliases_dir.sh
+if ! test -d ~/.bash_aliases.d/ || ! test -f ~/.bash_aliases.d/lazygit.sh || (test -f ~/.bash_aliases.d/lazygit.sh && ! grep -q "copy-to" ~/.bash_aliases.d/lazygit.sh); then
+    reade -Q "GREEN" -i "y" -p "Set up an alias so copy-to does a 'run all' when starting up lazygit? [Y/n]: " "y n" nstll
+    if [ "$nstll" == "y" ]; then
+        if ! test -f checks/check_aliases_dir.sh; then
+            eval "$(curl -fsSL https://raw.githubusercontent.com/excited-bore/dotfiles/main/checks/check_aliases_dir.sh)" 
+        else
+           . ./checks/check_aliases_dir.sh
+        fi
+        if ! test -f ~/.bash_aliases.d/lazygit.sh; then 
+            printf "alias lazygit=\"copy-to run all; lazygit\"\n" > ~/.bash_aliases.d/lazygit.sh
+            echo "$(tput setaf 2)File in ~/.bash_aliases.d/lazygit.sh"
+        fi
     fi
-    if ! test -f ~/.bash_aliases.d/lazygit.sh; then 
-        printf "alias lazygit=\"copy-to run all; lazygit\"\n" > ~/.bash_aliases.d/lazygit.sh
-        echo "File in ~/.bash_aliases.d/lazygit.sh"
-    fi
+    unset nstll
 fi
-unset nstll
